\name{bus_stop_milano}
\alias{bus_stop_milano}
\docType{data}
\title{
Milano bus stop
}
\description{
An extraction using the function poi.osm.
}
\usage{data("bus_stop_milano")}
\format{
  The format is:
List of 8
 $ bbox             : chr "45.3867381,9.0408867,45.5358482,9.2781103"
 $ overpass_call    : chr "[out:xml][timeout:25];\n(\n node  [\"highway\"=\"bus_stop\"] (45.3867381,9.0408867,45.5358482,9.2781103);\n way"| __truncated__
 $ meta             :List of 3
  ..$ timestamp       : chr "[ mer 4 mar 2020 18:07:02 ]"
  ..$ OSM_version     : chr "0.6"
  ..$ overpass_version: chr "Overpass API 0.7.55.7 8b86ff77"
 $ osm_points       :Formal class 'SpatialPointsDataFrame' [package "sp"] with 5 slots
  .. ..@ data       :'data.frame':	3038 obs. of  66 variables:
  .. .. ..$ osm_id                               : Factor w/ 3038 levels "1003229677","1003229681",..: 678 683 2977 2978 2979 2980 52 107 108 109 ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ name                                 : Factor w/ 1571 levels "\303\227","1 - Ingresso Ovest",..: 255 1393 1514 947 101 320 316 936 1513 354 ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ addr:postcode                        : Factor w/ 1 level "20094": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ alt_name                             : Factor w/ 21 levels "Foro Buonaparte dopo via Cusani",..: NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ amenity                              : Factor w/ 1 level "waste_basket": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ bench                                : Factor w/ 3 levels "no","yes","yes;no": 2 1 NA NA NA 1 NA 2 NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ building                             : Factor w/ 1 level "yes": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ bus                                  : Factor w/ 2 levels "no","yes": NA 2 NA NA NA NA NA NA NA 2 ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ bus_routes                           : Factor w/ 3 levels "51","708","Flixbus 408, 420, 426, 464, 494, 518, 519, 527, 528, 529, 544, 1420, 1426, 1590, N420, N424, N426, N494, N525, "| __truncated__: NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ comment                              : Factor w/ 11 levels "bus 94 does stop here again from 25 May 2016",..: NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ comment:en                           : Factor w/ 7 levels "bus 60 does not stop here",..: NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ comment:en_us                        : Factor w/ 1 level "end of line": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ comment:it                           : Factor w/ 72 levels "anche fermata sostituita M3 ed NM3 in direzione San Donato",..: NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ construction                         : Factor w/ 1 level "bus_stop": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ coommen:it                           : Factor w/ 1 level "senza indicatore tempo di attesa (predisposta, ma non installato)": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ covered                              : Factor w/ 3 levels "no","partial",..: NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ departures_board                     : Factor w/ 1 level "yes": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ description                          : Factor w/ 4 levels "54 > Duomo M1 M3\n73 > Duomo M1 M3\n84 > C.so di P.ta Vittoria",..: NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ disused                              : Factor w/ 1 level "yes": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ disused:highway                      : Factor w/ 1 level "bus_stop": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ disused:name                         : Factor w/ 1 level "Via Gioia - Via Cagliero": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ disused:passenger_information_display: Factor w/ 1 level "yes": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ disused:railway                      : Factor w/ 1 level "tram_stop": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ disused:ref                          : Factor w/ 1 level "12234": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ ele                                  : Factor w/ 1 level "87.60000000": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ end_date                             : Factor w/ 1 level "2014-03-22": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ fixme                                : Factor w/ 48 levels "a questa fermata fermano anche altre linee, non solo la 53",..: NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ highway                              : Factor w/ 1 level "bus_stop": 1 1 1 1 1 1 1 1 1 1 ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ image                                : Factor w/ 1 level "http://storage1.openstreetcam.org/files/photo/2016/6/15/lth/7455_a45e4_5761730f75ac1.jpg": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ in_use                               : Factor w/ 1 level "no": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ layer                                : Factor w/ 1 level "-1": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ lit                                  : Factor w/ 1 level "yes": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ loc_ref                              : Factor w/ 1 level "no": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ local_ref                            : Factor w/ 124 levels "10213","10214",..: NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ location                             : Factor w/ 2 levels "Largo Valera",..: NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ mapillary                            : Factor w/ 162 levels "_PuKMT6_d0lq_SXNdWDCBg",..: NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ name:de                              : Factor w/ 1 level "amor de sicilia": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ name:en                              : Factor w/ 1 level "Malpensa Airport-Centrale BUS": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ name:it                              : Factor w/ 1 level "Fondazione Prada": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ network                              : Factor w/ 11 levels "321","328","352, 352/",..: NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ note                                 : Factor w/ 16 levels "965 ferma qui",..: NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ note:it                              : Factor w/ 9 levels "con fermata anche di sostitutiva M1 ed NM1",..: NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ old_name                             : Factor w/ 134 levels "Boccaccio","Corso Europa - Piazza San Babila",..: NA NA NA 120 NA NA 116 NA 118 NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ operator                             : Factor w/ 16 levels "2 Busparkpl\303\244tze",..: NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ passenger_information_display        : Factor w/ 2 levels "no","yes": NA 1 NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ pole                                 : Factor w/ 2 levels "no","yes": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ public_transport                     : Factor w/ 2 levels "platform","stop_position": NA 1 NA NA NA NA NA NA NA 2 ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ railway                              : Factor w/ 1 level "tram_stop": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ ref                                  : Factor w/ 1770 levels "1","100057","10018",..: 350 652 NA NA 578 558 1113 353 NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ ref_route                            : Factor w/ 1 level "N6": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ route_ref                            : Factor w/ 90 levels "?","1","12175",..: NA NA NA NA 46 NA 71 NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ route_ref_1                          : Factor w/ 2 levels "47","98": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ route_ref_2                          : Factor w/ 1 level "47": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ shelter                              : Factor w/ 3 levels "no","yes","yes;no": 2 1 1 NA 2 1 2 2 NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ source                               : Factor w/ 2 levels "Bing","survey": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ surface                              : Factor w/ 1 level "asphalt": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ survey:date                          : Factor w/ 73 levels "2016-01-31","2016-05-20",..: NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ tactile_paving                       : Factor w/ 2 levels "no","yes": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ towards                              : Factor w/ 59 levels "Arese Stabile Alfa Romeo",..: NA NA NA NA 59 NA 30 NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ tram                                 : Factor w/ 2 levels "no","yes": NA 2 NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ trolleybus                           : Factor w/ 1 level "yes": NA NA NA NA NA NA NA NA NA 1 ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ type                                 : Factor w/ 1 level "public_transport": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ url                                  : Factor w/ 2 levels "http://orari.atm-mi.it/352_13735.pdf",..: NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ url_1                                : Factor w/ 1 level "http://orari.atm-mi.it/321_13741.pdf": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ url_2                                : Factor w/ 1 level "http://orari.atm-mi.it/321_13735.pdf": NA NA NA NA NA NA NA NA NA NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. ..$ wheelchair                           : Factor w/ 3 levels "limited","no",..: 3 NA NA NA NA NA 3 NA 3 NA ...
  .. .. .. ..- attr(*, "names")= chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. ..@ coords.nrs : num(0) 
  .. ..@ coords     : num [1:3038, 1:2] 9.14 9.17 9.15 9.16 9.16 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:3038] "21226899" "21298371" "95760548" "95760550" ...
  .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. ..@ bbox       : num [1:2, 1:2] 9.04 45.39 9.28 45.54
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:2] "x" "y"
  .. .. .. ..$ : chr [1:2] "min" "max"
  .. ..@ proj4string:Formal class 'CRS' [package "sp"] with 1 slot
  .. .. .. ..@ projargs: chr "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0"
 $ osm_lines        :Formal class 'SpatialLinesDataFrame' [package "sp"] with 4 slots
  .. ..@ data       :'data.frame':	0 obs. of  0 variables
  .. ..@ lines      : Named list()
  .. ..@ bbox       : num [1:2, 1:2] 9.04 45.39 9.28 45.54
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:2] "x" "y"
  .. .. .. ..$ : chr [1:2] "min" "max"
  .. ..@ proj4string:Formal class 'CRS' [package "sp"] with 1 slot
  .. .. .. ..@ projargs: chr "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0"
 $ osm_polygons     :Formal class 'SpatialPolygonsDataFrame' [package "sp"] with 5 slots
  .. ..@ data       :'data.frame':	1 obs. of  2 variables:
  .. .. ..$ highway: Factor w/ 1 level "bus_stop": 1
  .. .. ..$ surface: Factor w/ 1 level "asphalt": 1
  .. ..@ polygons   :List of 1
  .. .. ..$ 202420306:Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. .. ..@ Polygons :List of 1
  .. .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. .. ..@ labpt  : num [1:2] 9.08 45.45
  .. .. .. .. .. .. .. ..@ area   : num 2.18e-09
  .. .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. .. ..@ coords : num [1:5, 1:2] 9.08 9.08 9.08 9.08 9.08 ...
  .. .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. .. ..$ : chr [1:5] "2123875954" "2123875953" "2123875951" "2123875952" ...
  .. .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. .. ..@ plotOrder: int 1
  .. .. .. .. ..@ labpt    : num [1:2] 9.08 45.45
  .. .. .. .. ..@ ID       : num 2.02e+08
  .. .. .. .. ..@ area     : num 2.18e-09
  .. ..@ plotOrder  : int 1
  .. ..@ bbox       : num [1:2, 1:2] 9.04 45.39 9.28 45.54
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:2] "x" "y"
  .. .. .. ..$ : chr [1:2] "min" "max"
  .. ..@ proj4string:Formal class 'CRS' [package "sp"] with 1 slot
  .. .. .. ..@ projargs: chr "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0"
 $ osm_multilines   :Formal class 'SpatialLinesDataFrame' [package "sp"] with 4 slots
  .. ..@ data       :'data.frame':	0 obs. of  0 variables
Formal class 'data.frame' [package "methods"] with 4 slots
  .. .. .. ..@ .Data    : list()
  .. .. .. ..@ names    : chr(0) 
  .. .. .. ..@ row.names: int(0) 
  .. .. .. ..@ .S3Class : chr "data.frame"
  .. ..@ lines      : Named list()
  .. ..@ bbox       : num [1:2, 1:2] 9.04 45.39 9.28 45.54
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:2] "x" "y"
  .. .. .. ..$ : chr [1:2] "min" "max"
  .. ..@ proj4string:Formal class 'CRS' [package "sp"] with 1 slot
  .. .. .. ..@ projargs: chr "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0"
 $ osm_multipolygons:Formal class 'SpatialPolygonsDataFrame' [package "sp"] with 5 slots
  .. ..@ data       :'data.frame':	0 obs. of  0 variables
Formal class 'data.frame' [package "methods"] with 4 slots
  .. .. .. ..@ .Data    : list()
  .. .. .. ..@ names    : chr(0) 
  .. .. .. ..@ row.names: int(0) 
  .. .. .. ..@ .S3Class : chr "data.frame"
  .. ..@ polygons   : Named list()
  .. ..@ plotOrder  : int(0) 
  .. ..@ bbox       : num [1:2, 1:2] 9.04 45.39 9.28 45.54
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:2] "x" "y"
  .. .. .. ..$ : chr [1:2] "min" "max"
  .. ..@ proj4string:Formal class 'CRS' [package "sp"] with 1 slot
  .. .. .. ..@ projargs: chr "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0"
 - attr(*, "class")= chr [1:3] "list" "osmdata" "osmdata_sp"
}
\source{
Openstreetmap.org
}
\keyword{datasets}
