\name{magicwrap-package}
\alias{magicwrap-package}
\alias{magicwrap}
\docType{package}
\title{
\packageTitle{magicwrap}
}
\description{
\packageDescription{magicwrap}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{magicwrap}
\packageIndices{magicwrap}
A simple wrapper for OSM Overpass API and few utils functions.
At the moment the library has four functions: a wrapper to download data with OSM API, a function that turns polygons into points and two converter.
We used this functions in the past years to download bus stops from OSM and import it to DB.
}
\author{
\packageAuthor{magicwrap}

Maintainer: \packageMaintainer{magicwrap}
}
